ISSUE #36: Platform split (#50)
* WIP: Phase 2 compile

* Phase 3 done

* ISSUE #36: Re-establish a platform split / module setup

* Removing duplicate files

* Update windows instructions

* Removed outdated alias

* Working through a stub JVM impl

* Massage tests to work on JS + JVM

* More modules cross compile

* Fixing the tests

* Diagramming the repo

* Updated diagram tools

* Rationalising module relationships

* Diagram update

* Make indigo-extras tests platform agnostic

* Mutable KVP and Batch instances

* Scalafix / Scalafmt

* Tweaked Claude instructions

* ISSUE #11: Replacing old entry points with 'next' versions

Removed left over SceneViewModel reference

Reorganise CaptureScreenScene model to improve clarity

Added namespace to SceneBlenderShader reference

Changed name to Game.ShaderPlayground

* Post-rebase fix-ups

* Fixed component group tests

* Updated diagram

* Remove Indigo Next README.md

+++ indigo-core/test/src/indigo/core/temporal/SignalReaderTests.scala
@@ -53,7 +53,7 @@ class SignalReaderTests extends munit.FunSuite {
     val actual: SignalReader[String, (String, Int)] =
       for {
         a <- SignalReader.fromSignal(Signal.Time)
-        b <- SignalReader((msg: String) => Signal.fixed(msg + a.toString()))
+        b <- SignalReader((msg: String) => Signal.fixed(msg + a.toLong.toString()))
         c <- SignalReader.fixed((b, b.length))
       } yield c
